+++ src/Nethermind/Nethermind.Specs/TestSpecProvider.cs
@@ -1,4 +1,4 @@
-//  Copyright (c) 2018 Demerzel Solutions Limited
+//  Copyright (c) 2021 Demerzel Solutions Limited
 //  This file is part of the Nethermind library.
 // 
 //  The Nethermind library is free software: you can redistribute it and/or modify
@@ -36,7 +36,7 @@ namespace Nethermind.Specs
         public IReleaseSpec SpecToReturn { get; set; }
 
         public long? DaoBlockNumber { get; set; }
-        public int ChainId { get; set; }
+        public long ChainId { get; set; }
         public long[] TransitionBlocks { get; set; } = new long[] {0};
     }
 }
